Update 1Password, secret spec
Instead of using separate 1Password entries per secret, use one
entry per application and make use of the complete customization
of the Server type so that all the key and value pairs can be the
static secrets for that application.

Document special handling for pull-secret and improve the
documentation for the Vault read secret.

Improve the secrets.yaml specification to allow the generate and
copy stanzas to also be conditional.